Refactor ChapterPublisher
